:recycle: Rename to CityId
